Initial extraction of components from 3 real-world components complete. This still needs:
1) generators
Some stuff is absent because it was previously created by hand and is
bes.t done as a generator now that we are a gem.

2) tests
Some stuff is still in need of tests.  Stuff was previously tested
in_memoryes/codesherpas/Stone-Path/scratch/stonepath/README the apps it
was part of; now the little sample app in test needs to have full tests.

3) docs
The major strength in this stuff isn't the code - it is the modeling
commitVolumes/codesherpas/Stone-Path/scratch/stonepath/READMEncepts
supported by the code.  I have a lot of writing to do...